requirements.txt
Pillow>=9.0.0
matplotlib>=3.4.0
numpy>=2.2.6